+++ source/detection_block/BlockRecognizer.py
@@ -174,58 +174,6 @@ class BlockExtractor():
         # 不要な情報が黒色になっているので白色に変換する
         return cv2.addWeighted(dst, 1, cv2.bitwise_not(mask), 1, 0)
 
-def test_result():
-    circles_coordinates = {
-        'c00': (34, 61), 'c10': (210, 61), 'c20': (393, 56), 'c30': (573, 58),
-        "b1": (122, 153), "b2": (302, 155), "b3": (480, 150),
-        'c01': (36, 243), 'c11': (217, 242), 'c21': (392, 243), 'c31': (577, 241),
-        "b4": (127, 338), "b5": (480, 339),
-        'c02': (43, 425), 'c12': (219, 429), 'c22': (399, 426), 'c32': (572, 427),
-        "b6": (130, 521), "b7": (307, 521), "b8": (481, 520),
-        'c03': (49, 608), 'c13': (227, 606), 'c23': (400, 607), 'c33': (578, 608)
-        }
-    recognizer = BlockRecognizer()
-    img = cv2.imread('result.png')
-    black, color, cc = recognizer.recognize(img, circles_coordinates)
-
-    result_cc_blocks = (
-        (Color.BLUE, Color.WHITE, Color.BLUE, Color.WHITE),
-        (Color.WHITE, Color.GREEN, Color.WHITE, Color.YELLOW),
-        (Color.RED, Color.WHITE, Color.YELLOW, Color.WHITE),
-        (Color.WHITE, Color.BLACK, Color.WHITE, Color.GREEN)
-        )
-    
-    assert (5, 7) == (black, color)
-    for row in range(4):
-        for col in range(4):
-            assert cc.cross_circles[col][row] == result_cc_blocks[col][row]
-    
-
-def test_result1():
-    circles_coordinates = {
-        'c00': (43, 48), 'c10': (217, 50), 'c20': (398, 47), 'c30': (576, 48),
-        "b1": (130, 145), "b2": (304, 148), "b3": (487, 147),
-        'c01': (42, 232), 'c11': (222, 237), 'c21': (396, 236), 'c31': (578, 235),
-        "b4": (131, 331), "b5": (484, 333),
-        'c02': (46, 420), 'c12': (222, 422), 'c22': (402, 418), 'c32': (571, 419),
-        "b6": (133, 517), "b7": (311, 517), "b8": (483, 515),
-        'c03': (52, 606), 'c13': (231, 604), 'c23': (399, 608), 'c33': (581, 607)}
-    recognizer = BlockRecognizer()
-    img = cv2.imread('result1.png')
-    black, color, cc = recognizer.recognize(img, circles_coordinates)
-    
-    # result1_cc_blocks[col][row]
-    result1_cc_blocks = (
-        (Color.BLUE, Color.WHITE, Color.BLUE, Color.WHITE),
-        (Color.WHITE, Color.YELLOW, Color.WHITE, Color.YELLOW),
-        (Color.RED, Color.WHITE, Color.RED, Color.WHITE),
-        (Color.WHITE, Color.BLACK, Color.WHITE, Color.GREEN)
-        )
-    
-    assert (5, 2) == (black, color)
-    for row in range(4):
-        for col in range(4):
-            assert cc.cross_circles[col][row] == result1_cc_blocks[col][row]
 
 if __name__ == '__main__':
     pass
